chore: use double instead of float

+++ benchmarks/OklabBenchmark.cpp
@@ -5,7 +5,7 @@ static void BM_InterpolateColor(benchmark::State &state)
 {
     RGB color1 = {255, 0, 0}; // Red
     RGB color2 = {0, 0, 255}; // Blue
-    double t = 0.5;
+    float t = 0.5;
     for (auto _ : state)
     {
         RGB result = interpolateColor(color1, color2, t);
@@ -20,7 +20,7 @@ static void BM_InterpolateRandomColors(benchmark::State &state)
     {
         RGB color1 = {rand() % 256, rand() % 256, rand() % 256};
         RGB color2 = {rand() % 256, rand() % 256, rand() % 256};
-        double t = static_cast<double>(rand()) / RAND_MAX;
+        float t = static_cast<float>(rand()) / RAND_MAX;
         RGB result = interpolateColor(color1, color2, t);
         benchmark::DoNotOptimize(result);
     }
